+++ src/main/java/com/example/sportstore06/repository/IBillRepository.java
@@ -40,85 +40,85 @@ public interface IBillRepository extends JpaRepository<Bill, Integer> {
     // thong ke
     //admin
     @Query("SELECT COALESCE(SUM(b.total), 0) FROM Bill b")
-    double getAllTotal();
+    Long getAllTotal();
 
     @Query("SELECT COALESCE(SUM(b.total), 0) FROM Bill b " +
             "WHERE b.updated_at >= :startDate AND b.updated_at <= :endDate")
-    double getAllTotal(
+    Long getAllTotal(
             @Param("startDate") Timestamp startDate,
             @Param("endDate") Timestamp endDate);
 
     @Query("SELECT COUNT(b) FROM Bill b")
-    double getAllCount();
+    Long getAllCount();
 
     @Query("SELECT COUNT(b) FROM Bill b " +
             "WHERE b.updated_at >= :startDate AND b.updated_at <= :endDate")
-    double getAllCount(@Param("startDate") Timestamp startDate,
+    Long getAllCount(@Param("startDate") Timestamp startDate,
                        @Param("endDate") Timestamp endDate);
 
     // add state
 
     @Query("SELECT COALESCE(SUM(b.total), 0) FROM Bill b WHERE b.state =:state")
-    double getAllTotal(@Param("state") Integer state);
+    Long getAllTotal(@Param("state") Integer state);
 
     @Query("SELECT COALESCE(SUM(b.total), 0) FROM Bill b " +
             "WHERE b.updated_at >= :startDate AND b.updated_at <= :endDate AND b.state =:state")
-    double getAllTotal(
+    Long getAllTotal(
             @Param("startDate") Timestamp startDate,
             @Param("endDate") Timestamp endDate,
             @Param("state") Integer state);
 
     @Query("SELECT COUNT(b) FROM Bill b WHERE b.state =:state")
-    double getAllCount(@Param("state") Integer state);
+    Long getAllCount(@Param("state") Integer state);
 
     @Query("SELECT COUNT(b) FROM Bill b " +
             "WHERE b.updated_at >= :startDate AND b.updated_at <= :endDate AND b.state =:state")
-    double getAllCount(@Param("startDate") Timestamp startDate,
+    Long getAllCount(@Param("startDate") Timestamp startDate,
                        @Param("endDate") Timestamp endDate,
                        @Param("state") Integer state);
 
     //business
 
     @Query("SELECT COALESCE(SUM(b.total), 0) FROM Bill b WHERE  b.id_business = :idBusiness")
-    double getAllTotalBusiness(@Param("idBusiness") Integer idBusiness);
+    Long getAllTotalBusiness(@Param("idBusiness") Integer idBusiness);
 
     @Query("SELECT COALESCE(SUM(b.total), 0) FROM Bill b " +
             "WHERE b.updated_at >= :startDate AND b.updated_at <= :endDate AND b.id_business = :idBusiness")
-    double getAllTotalBusiness(
+    Long getAllTotalBusiness(
             @Param("startDate") Timestamp startDate,
             @Param("endDate") Timestamp endDate,
             @Param("idBusiness") Integer idBusiness);
 
     @Query("SELECT COUNT(b) FROM Bill b WHERE b.id_business = :idBusiness")
-    double getAllCountBusiness(@Param("idBusiness") Integer idBusiness);
+    Long getAllCountBusiness(@Param("idBusiness") Integer idBusiness);
 
     @Query("SELECT COUNT(b) FROM Bill b " +
             "WHERE b.updated_at >= :startDate AND b.updated_at <= :endDate AND b.id_business = :idBusiness")
-    double getAllCountBusiness(@Param("startDate") Timestamp startDate,
+    Long getAllCountBusiness(@Param("startDate") Timestamp startDate,
                        @Param("endDate") Timestamp endDate,
                        @Param("idBusiness") Integer idBusiness);
 
     // add state
 
     @Query("SELECT COALESCE(SUM(b.total), 0) FROM Bill b WHERE b.state =:state AND b.id_business = :idBusiness")
-    double getAllTotalBusiness(@Param("state") Integer state,
+    Long getAllTotalBusiness(@Param("state") Integer state,
                        @Param("idBusiness") Integer idBusiness);
 
     @Query("SELECT COALESCE(SUM(b.total), 0) FROM Bill b " +
             "WHERE b.updated_at >= :startDate AND b.updated_at <= :endDate AND b.state =:state AND b.id_business = :idBusiness")
-    double getAllTotalBusiness(
+    Long getAllTotalBusiness(
             @Param("startDate") Timestamp startDate,
             @Param("endDate") Timestamp endDate,
             @Param("state") Integer state,
             @Param("idBusiness") Integer idBusiness);
 
     @Query("SELECT COUNT(b) FROM Bill b WHERE b.state =:state AND b.id_business = :idBusiness")
-    double getAllCountBusiness(@Param("state") Integer state,
+    Long getAllCountBusiness(@Param("state") Integer state,
                        @Param("idBusiness") Integer idBusiness);
 
     @Query("SELECT COUNT(b) FROM Bill b " +
             "WHERE b.updated_at >= :startDate AND b.updated_at <= :endDate AND b.state =:state and b.id_business = :idBusiness")
-    double getAllCountBusiness(@Param("startDate") Timestamp startDate,
+    Long getAllCountBusiness(@Param("startDate") Timestamp startDate,
                        @Param("endDate") Timestamp endDate,
                        @Param("state") Integer state,
                        @Param("idBusiness") Integer idBusiness);
